Wait for completion of the first CoA registration
During first registration, when dynamic keying is used, it is a
bad idea to change our mind and select another CoA. This leads to
desync between local IKE daemon and remote one. We must wait
for the completion (or failure) of the first registration.

by Arnaud Ebalard <arno@natisbad.org> -- Under GPLv2.