Let the IKE keying complete for soft handovers
UMIP is not aware of the keying performed by an IKE daemon when one
is used to negotiate SA protecting signaling and traffic. In case of
a soft handover, there is no rush this is better to let the IKE daemon
complete its job instead of shooting it a bullet in the foot.

XXX This patch is a work in progress, i.e. for testing purposes.

by Arnaud Ebalard <arno@natisbad.org> -- Under GPLv2.