Split the BA processing code
mn_recv_ba(), the MH BA handler for MN is difficult to follow due to
its length but also because it includes the logic for both main kinds
of BUL entries: those sent to the HA (bule->flags & IP6_MH_BU_HOME
!= 0 ) and those sent to the CN (bule->flags & IP6_MH_BU_HOME == 0).

This patch splits the code performing the processing of the BA in
two helpers: mn_process_ha_ba() and mn_process_cn_ba() which are
called from a simplified handler (mn_recv_ba()). This function does
the option parsing and some initial check before dispatching the
work to one of those two handlers.

In the end, there is not that much duplicated code and this is far
clearer.

by Arnaud Ebalard <arno@natisbad.org> -- Under GPLv2.