Decide if a BU acceptation depending on DoRouteOptimizationCN
For a while now, the daemon man page documents DoRouteOptimizationCN
functionality: it is just said it does nothing at the moment. It also
documents BindingAclPolicy and DefaultBindingAclPolicy as HA specific
parameters which control binding of MN on the HA.

In the sources (cn.c), when a BU is received with H flags not set and
processed by cn_recv_bu() (for HA, MN or CN), default_discard_binding()
policy function is called, which performs a decision based on
BindingAclPolicy and DefaultBindingAclPolicy values provided in the
configuration file.

This is not consistent with the behavior advertised in the man.

The patch below does the following: it replaces the call to
default_discard_binding() in cn_recv_bu() to decide if the BU
should be accepted by a check of conf.DoRouteOptimizationCN value.

Note that this patch is a first step towards more granularity for
binding selection when acting as CN. This is provided by a companion
patch.

by Arnaud Ebalard <arno@natisbad.org> -- Under GPLv2.